Unify from_json array map tests into FromJsonToRawMapTest
Move the four MapValueType.ARRAY_OF_STRING tests and their helpers (STRUCT_TYPE, LIST_TYPE, pair, assertKeyAndStructNonNullable) from FromJsonToRawMapArrayTest into FromJsonToRawMapTest so all extractRawMapFromJsonString tests share one file and the common getOptions, and delete the now-redundant FromJsonToRawMapArrayTest.

---------
